% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRCRTwa-package.R
\docType{package}
\name{DRCRTwa-package}
\alias{DRCRTwa}
\alias{DRCRTwa-package}
\title{DRCRTwa: Doubly Robust While-Alive Estimation for Randomized Trials}
\description{
The \code{DRCRTwa} package implements compiled doubly robust estimation of
exposure-weighted while-alive recurrent-event rates for individually
randomized and cluster-randomized trials in the presence of a terminal event
and covariate-dependent censoring. The main fitting function is
\code{\link[=DRCRT_WA]{DRCRT_WA()}}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fancy575/DRCRTwa}
  \item Report bugs at \url{https://github.com/fancy575/DRCRTwa/issues}
}

}
\author{
\strong{Maintainer}: Xi Fang \email{xfang@mcw.edu}

Authors:
\itemize{
  \item Xi Fang \email{xfang@mcw.edu}
  \item Da Zhao \email{dzhao@yale.edu}
  \item Fan Li \email{fan.f.li@yale.edu}
}

}
\keyword{internal}
